+++ src/Web/Mailchimp/Common.hs
@@ -1,20 +1,19 @@
 ----------------------------------------------------------------------
 -- |
--- Module: Web.MailChimp.Common
+-- Module: Web.Mailchimp.Common
 -- Description:
 --
 --
 --
 ----------------------------------------------------------------------
 
-module Web.MailChimp.Common
+module Web.Mailchimp.Common
   ( Id
   )
   where
 
 -- text
 import Data.Text (Text)
-
 
 -- |
 --