+++ src/Web/Mailchimp/Key.hs
@@ -1,13 +1,13 @@
 ----------------------------------------------------------------------
 -- |
--- Module: Web.MailChimp.Key
+-- Module: Web.Mailchimp.Key
 -- Description:
 --
 --
 --
 ----------------------------------------------------------------------
 
-module Web.MailChimp.Key
+module Web.Mailchimp.Key
   ( DataCenter
   , Key
   , parseDataCenter
@@ -21,22 +21,19 @@ import Data.Attoparsec.ByteString.Char8 (anyChar, char)
 -- bytestring
 import Data.ByteString.Char8
 
-
 -- |
 --
--- A MailChimp data center.
+-- A Mailchimp data center.
 
 type DataCenter =
   ByteString
 
-
 -- |
 --
--- A MailChimp API key.
+-- A Mailchimp API key.
 
 type Key =
   ByteString
-
 
 -- |
 --
@@ -47,7 +44,6 @@ parseDataCenter
   -> Either String DataCenter
 parseDataCenter =
   parseOnly dataCenterParser
-
 
 -- |
 --