+++ src/Web/Mailchimp/List.hs
@@ -1,0 +1,23 @@
+----------------------------------------------------------------------
+-- |
+-- Module: Web.Mailchimp.List
+-- Description:
+--
+--
+--
+----------------------------------------------------------------------
+
+module Web.Mailchimp.List
+  ( ListId
+  )
+  where
+
+-- mailchimp
+import Web.Mailchimp.Common (Id)
+
+-- |
+--
+-- A list ID.
+
+type ListId =
+  Id