+++ src/Web/Mailchimp/Tutorial.hs
@@ -3,14 +3,14 @@
 
 ----------------------------------------------------------------------
 -- |
--- Module: Web.MailChimp.Tutorial
+-- Module: Web.Mailchimp.Tutorial
 -- Description:
 --
 --
 --
 ----------------------------------------------------------------------
 
-module Web.MailChimp.Tutorial
+module Web.Mailchimp.Tutorial
   ( example
   )
   where
@@ -22,11 +22,10 @@ import System.Environment (getEnv)
 import Data.ByteString.Char8 (pack)
 
 -- mailchimp
-import Web.MailChimp
+import Web.Mailchimp
 
 -- text
 import qualified Data.Text as Text
-
 
 example :: IO ()
 example = do
@@ -35,17 +34,15 @@ example = do
   listId <- fmap Text.pack (getEnv "MAILCHIMP_LIST_ID")
 
   let
-    AuthClient {..} = makeAuthClientWithKey key
-    ListMemberClient {..} = makeListMemberClient listId
-
+    basicAuthData = makeBasicAuthData key
   let
     member =
       (makeListMemberRequest "sd@sd.com" Pending)
         { listMemberMergeFields = [("FNAME", "Juan")]
         }
 
-  eitherAdd <- run manager key (addListMember member)
---  eitherAdd <- run manager key (getLinks (BasicAuthData "" key))
+  eitherAdd <- run manager key (addListMember listMemberClient basicAuthData listId member)
+  -- eitherAdd <- run manager key (getLinks mainClient basicAuthData)
 
   case eitherAdd of
     Left err ->