+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+# 0.4.0
+
+### Breaking changes
+
+* Upgrade to Servant >= 0.16
+* Switch from MailChimp to Mailchimp
+
+# 0.3.0
+
+- Added pagination to get list members and helper to get all list
+  members (7593fb3).
+
+- Fixed list members response (8f47755).
+
 # 0.2.1
 
 - Tested with GHC 8.0.2.